1.修复通讯录bug
git-svn-id: http://10.22.1.27:8000/svn/rooten/MSMS/src/client/IOS/IM@1307 47c12b7e-2652-5545-b168-31c46e2fce3f

+++ IM/os/tb/OsSql.m
@@ -43,7 +43,7 @@ NSString *kSqlOsItemsCreate = @"CREATE TABLE IF NOT EXISTS `tb_os_items` ("
                                         "`org`	TEXT,"
                                         "PRIMARY KEY(uid)"
                                         ");";
-NSString *kSqlOsItemsDel = @"DELETE FROM `tb_os_items`;";
+NSString *kSqlOsItemsDel = @"DELETE FROM `tb_os_items` WHERE `uid` = ?;";
 NSString *kSqlOsItemsQuery = @"SELECT * FROM `tb_os_items`;";
 NSString *kSqlOsItemsQueryByOrg = @"SELECT * FROM `tb_os_items` WHERE `org` = ?;";
 NSString *kSqlOsItemsQueryByUid = @"SELECT * FROM `tb_os_items` WHERE `uid` = ?;";
